Remove test for spatial distances in `check_data.IMMspatial`
I removed the check if the distances are in radians from the `check_data.IMMspatial` as I did not want to assume that everyone will use the IMM with distances around the circle.

Instead, I introduced a check that all distances need to be positive as negative distances would result in them getting higher activations then the target.

I also added a test to verify that this is working correctly.